PoPie: Add Analyzer tests
Analyzer is the most critical part of the PoPie. Having tests for the
rest would be nice, but it's not top priority.

With every new bug, at least one new test should be added in the same
commit. This will ensure that we can look back at example of a code
that we weren't handling correctly.